vmm: api: Move HttpError enum to http module
Minor rearrangement of code to make it easier to implement refactoring.
